requirements.txt
# Whisper 相關
faster-whisper
torch<2.6
torchaudio<2.6

# Speaker Diarization
pyannote.audio
huggingface-hub>=0.21,<0.22

# 音檔處理
pydub

# API 相關
fastapi
uvicorn[standard]
python-multipart

# 系統監控
psutil

# HTTP 客戶端
requests

# Google OAuth
google-auth

# LLM API
openai
google-generativeai
